Misc: Rename files
* `app.py` -> `api.py`
* `app_factory.py` -> `app.py`
